chore: small refactor + rename
