+++ internal/impl/parquet/util.go
@@ -2,13 +2,9 @@ package parquet
 
 import (
 	"encoding/json"
-	"fmt"
 	"math/rand"
 	"strings"
 	"time"
-
-	"github.com/parquet-go/parquet-go"
-	"github.com/warpstreamlabs/bento/internal/value"
 )
 
 func scrubJSONNumbers(v any) any {
@@ -40,42 +36,6 @@ func scrubJSONNumbersObj(obj map[string]any) {
 func scrubJSONNumbersArr(arr []any) {
 	for i, v := range arr {
 		arr[i] = scrubJSONNumbers(v)
-	}
-}
-
-func valueOf(val any, field parquet.Field) (any, error) {
-	if val == nil {
-		return nil, nil
-	}
-
-	if !field.Leaf() {
-		return val, nil
-	}
-
-	if arr, ok := val.([]any); ok {
-		for i, v := range arr {
-			res, err := valueOf(v, field)
-			if err != nil {
-				return nil, fmt.Errorf("index %d: %w", i, err)
-			}
-			arr[i] = res
-		}
-		return arr, nil
-	}
-
-	switch field.Type().Kind() {
-	case parquet.Int32:
-		return value.IToInt32(val)
-	case parquet.Int64:
-		return value.IToInt(val)
-	case parquet.Float:
-		return value.IToFloat32(val)
-	case parquet.Double:
-		return value.IToFloat64(val)
-	case parquet.Boolean:
-		return value.IToBool(val)
-	default:
-		return val, nil
 	}
 }
